+++ CHANGES.md
@@ -1,5 +1,11 @@
 # ChangeLog
+
+## v3.4.1
+
+* Added SIRF interfaces to STIR's `get_STIR_doc_dir()` and `get_STIR_examples_dir()`.
+
 ## v3.4.0
+
 * changed Python test framework to pytest
 * MR:
   - Added support for spiral trajectories that are pre-computed by the user by exposing a setter.